requirements.txt
selenium
requests
colorama
customtkinter
Pillow
psutil